+++ philo/src/utils/get_ms_timeval.c
@@ -6,7 +6,7 @@
 /*   By: mgo <mgo@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 13:21:39 by mgo               #+#    #+#             */
-/*   Updated: 2022/03/19 12:43:24 by mgo              ###   ########.fr       */
+/*   Updated: 2022/03/19 15:02:57 by mgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ long long	get_ms_timeval(struct timeval tv)
 	return (ret);
 }
 
-long long	get_time_ms(void)
+long int	get_time_ms(void)
 {
-	long long		ret;
+	long int		ret;
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
